set nodeUrl as parameter connection

+++ crossboarder-app/src/main/java/us/lacchain/crossborder/management/repository/MovementRepository.java
@@ -25,7 +25,7 @@ public interface MovementRepository extends JpaRepository<Movement, String> {
     @Transactional
     @Modifying(clearAutomatically = true)
     @Query(value = "update movements set status = 4, received_amount = :amount, operation_executed = :transactionHash where id = :operationId", nativeQuery=true)
-    void setTransferExecuted(@Param("transactionHash") String transactionHash,@Param("operationId") String operationId, @Param("amount") int amount);
+    void setTransferExecuted(@Param("transactionHash") String transactionHash,@Param("operationId") String operationId, @Param("amount") float amount);
 
     @Transactional
     @Modifying(clearAutomatically = true)
